# backend/src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Seax 工作流程與組織記憶 API",
    "version": "0.2.0",
    "description": "GlobalMemory schemaVersion 2：版本化知識、結案案例、scope、證據、UNKNOWN 原因及回饋診斷。POC 沿用 /api/v1；既有外部消費者需同步升級。必填文字不得只有空白。所有 POST 須 Idempotency-Key。Job、交易、重試及結案規則沿用 backend/spec.md；AI_INPUT_TOO_LARGE 不可重試，不以 UNKNOWN 掩蓋技術失敗。"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/global-memory": {
      "post": {
        "summary": "初始化全系統共用的組織架構與部門職能",
        "operationId": "postGlobalMemory_global-memory",
        "tags": [
          "global-memory"
        ],
        "responses": {
          "201": {
            "description": "初始化全系統共用的組織架構與部門職能",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalMemory"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "description": "永久冪等 key；相同路徑/key/body 回放原成功回應，不同 body 回傳 409。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializeMemory"
              }
            }
          }
        }
      },
      "get": {
        "summary": "查詢最新或指定版本的組織表",
        "operationId": "getGlobalMemory_global-memory",
        "tags": [
          "global-memory"
        ],
        "responses": {
          "200": {
            "description": "查詢最新或指定版本的組織表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalMemory"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false
          }
        ]
      }
    },
...
    "/projects": {
      "post": {
        "summary": "建立專案、提交需求並啟動首次 AI 分析",
        "operationId": "postCreateProjectResult_projects",
        "tags": [
          "projects"
        ],
        "responses": {
          "202": {
            "description": "建立專案、提交需求並啟動首次 AI 分析",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectResult"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "description": "永久冪等 key；相同路徑/key/body 回放原成功回應，不同 body 回傳 409。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          }
        }
      },
      "get": {
        "summary": "查詢專案列表",
        "operationId": "getProjectPage_projects",
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "查詢專案列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPage"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "required": false
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "CLOSED"
              ]
            },
            "required": false
          }
        ]
      }
    },
    "/projects/{projectId}": {
      "get": {
        "summary": "查詢專案詳情、需求內容與任務關聯",
        "operationId": "getProjectDetails_projects_projectId",
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "查詢專案詳情、需求內容與任務關聯",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetails"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ]
      }
    },
    "/projects/{projectId}/report": {
      "get": {
        "summary": "查詢報告、工作相依順序及歸屬結果",
        "operationId": "getReport_projects_projectId_report",
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "查詢報告、工作相依順序及歸屬結果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "description": "以 DAG 拓撲排序回傳工作；每一步候選工作依 createdAt、id 升冪。相依 ID 陣列按 UUID 升冪。AI 指派的 assignmentReason 會隨 workflow 回傳。"
      }
    },
    "/projects/{projectId}/workflows": {
      "post": {
        "summary": "在指定專案中手動新增工作任務及前置相依",
        "operationId": "postWorkflowResult_projects_projectId_workflows",
        "tags": [
          "projects"
        ],
        "responses": {
          "201": {
            "description": "在指定專案中手動新增工作任務及前置相依",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResult"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "description": "永久冪等 key；相同路徑/key/body 回放原成功回應，不同 body 回傳 409。"
          }
        ],
        "description": "省略歸屬欄位時新增為 UNASSIGNED、修改時保留；只傳 departmentId 則有效 UUID 為 ASSIGNED、null 為 UNASSIGNED。明確 ASSIGNED 須提供 departmentId；UNKNOWN/UNASSIGNED 的 departmentId 須省略或 null。手動判定來源由後端設為 USER，並清除既有 AI assignmentReason。相依須同 Report、無重複、自我或循環。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflow"
              }
            }
          }
        }
      }
    },
    "/workflows/{workflowId}": {
      "patch": {
        "summary": "修改工作任務內容、前置相依或部門歸屬",
        "operationId": "patchWorkflowResult_workflows_workflowId",
        "tags": [
          "workflows"
        ],
        "responses": {
          "200": {
            "description": "修改工作任務內容、前置相依或部門歸屬",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResult"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workflowId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "description": "省略歸屬欄位時新增為 UNASSIGNED、修改時保留；只傳 departmentId 則有效 UUID 為 ASSIGNED、null 為 UNASSIGNED。明確 ASSIGNED 須提供 departmentId；UNKNOWN/UNASSIGNED 的 departmentId 須省略或 null。手動判定來源由後端設為 USER，並清除既有 AI assignmentReason。相依須同 Report、無重複、自我或循環。 相依省略保留、[] 清空，null 不合法；僅重排不產生差異。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchWorkflow"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "刪除工作任務並保留變更紀錄",
        "operationId": "deleteDeleteWorkflowResult_workflows_workflowId",
        "tags": [
          "workflows"
        ],
        "responses": {
          "200": {
            "description": "刪除工作任務並保留變更紀錄",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkflowResult"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "expectedReportVersion",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "required": true
          },
          {
            "name": "reason",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2000
            },
            "required": false
          },
          {
            "name": "workflowId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "description": "有後續工作直接依賴時回傳 WORKFLOW_HAS_DEPENDENTS，details.dependentWorkflowIds 列出 ID。"
      }
    },
    "/projects/{projectId}/analyses": {
      "post": {
        "summary": "啟動全部或僅未歸屬項目的 AI 歸屬分析",
        "operationId": "postAnalysisResult_projects_projectId_analyses",
        "tags": [
          "projects"
        ],
        "responses": {
          "202": {
            "description": "啟動全部或僅未歸屬項目的 AI 歸屬分析",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "description": "永久冪等 key；相同路徑/key/body 回放原成功回應，不同 body 回傳 409。"
          }
        ],
        "description": "ALL_REANALYZE 接受時立即清除全部人工/AI 歸屬與 UNKNOWN；UNASSIGNED_ANALYZE 涵蓋 UNASSIGNED 及 UNKNOWN。保留名稱、描述與相依。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          }
        }
      }
    },
    "/projects/{projectId}/report-diffs": {
      "get": {
        "summary": "查詢報告的變更歷程",
        "operationId": "getReportDiffPage_projects_projectId_report-diffs",
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "查詢報告的變更歷程",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDiffPage"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "required": false
          },
          {
            "name": "projectId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GlobalMemory": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "const": 2
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "source": {
            "type": "string",
            "enum": [
              "INITIAL",
              "FEEDBACK"
            ]
          },
          "sourceProjectId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "departments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Department"
            }
          },
          "relationshipsDescription": {
            "type": "string",
            "minLength": 0,
            "maxLength": 100000
          },
          "counts": {
            "$ref": "#/components/schemas/MemoryCounts"
          }
        },
        "required": [
          "schemaVersion",
          "version",
          "source",
          "sourceProjectId",
          "createdAt",
          "departments",
          "relationshipsDescription",
          "counts"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "CreateProjectResult": {
        "type": "object",
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "report": {
            "$ref": "#/components/schemas/Report"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "required": [
          "project",
          "report",
          "job"
        ],
        "additionalProperties": false
      },
...
    }
  }
}